Move client.tsx to index.
